% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_poll_by_id.R
\name{get_poll_by_id}
\alias{get_poll_by_id}
\title{Get Poll Data from Tweet ID}
\arguments{
\item{tweet_id}{character, the id of the tweet with poll to query}
}
\value{
list
}
\description{
Return a data.frame of poll data from Twitter given a poll ID.
The output data.frame will contain five columns: (1) poll_question,
(2) poll_options, (3) total_responses, (4) duration, (5) date, (6) user.
}
\examples{
\dontrun{
  get_poll_by_id('4235234')
}
}
