% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_polls_from_user.R
\name{get_polls_from_user}
\alias{get_polls_from_user}
\title{Get Polls from User}
\arguments{
\item{username}{character, the username of the twitter user to query}
}
\value{
data.frame
}
\description{
Return a data.frame of poll ids for a given Twitter user.
REQUIRES: Credentials for the Twitter API need to be stored as environment
variables since this function calls directly to the bash_profile.
}
\examples{
\dontrun{
  get_polls_from_user('PollzOnTwitta')
}
}
