% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_poll.R
\name{visualize_poll}
\alias{visualize_poll}
\title{Visualize Poll Data}
\arguments{
\item{poll_obj}{A list containing poll responses from get_poll_by_id()}

\item{show_user}{A logical to display user}

\item{show_duration}{A logical to display duration of the poll}

\item{show_date}{A logical to display end date of the poll}
}
\value{
A bar chart containing poll responses
including additional information if at least one of
\itemize{
\item show_user
\item show_duration
\item show_date
was set to TRUE
}
}
\description{
Return a bar chart of poll responses.
Option to include additional information in the text box
if at least one of (1) show_user, (2) show_duration, and
(3)show_date is set to True.
}
\examples{
poll_obj <- list()
poll_obj['text'] <- 'How normal is it to shower before the gym'
poll_obj['duration'] <- as.integer(1440)
poll_obj['date'] <- '2022-01-12T23:08:22.000Z'
poll_obj['poll options'] <- list(data.frame(
'position' = as.integer(c(1, 2, 3, 4)),
'label' = c('Not normal', 'Completely wack', 'Insane', 'Run'),
'votes' = as.integer(c(295, 124, 146, 50))))
poll_obj['user'] <- 'RadaNotSay'
poll_obj['total'] <- as.integer(615)
visualize_poll(poll_obj, show_duration=TRUE)
}
